requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
motor>=3.3.0
pymongo>=4.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.0.0
httpx>=0.25.0
aiohttp